Fix limit formatting.

+++ cmd/frontend/graphqlbackend/lsif_dumps.go
@@ -87,7 +87,7 @@ func (r *lsifDumpConnectionResolver) compute(ctx context.Context) ([]*types.LSIF
 			query.Set("query", *r.opt.Query)
 		}
 		if r.opt.Limit != nil {
-			query.Set("limit", strconv.FormatInt(*r.opt.Limit, 10))
+			query.Set("limit", strconv.FormatInt(int64(*r.opt.Limit), 10))
 		}
 
 		resp, err := lsif.BuildAndTraceRequest(ctx, path, query)
